ARM: dts: ts7180: Add ts7100-wdt (watchdog) on i2c1
This is always needed, because otherwise a timeout set in U-Boot will
not be cancelled and the board will reboot.

The entry's comment includes a breadcrumb to explain why the timeout
that is read back from hardware is likely to differ from whatever is
set here:

Userspace procs can still start feeding with different values later -
watchdog(8) defaults to 60 seconds.

The CPU's builtin watchdog (wdog1) has a trace enabling it to assert
RST# (analogous to EXT_RESET#/PWR_CYCLE# on the C99). We disable the
watchdog, but leave the pinmux settings in place.

wdog1 should be disabled, because otherwise both it and the ts7100-wdt
will have a 60-second timeout, making it possibly random which
reset circuitry will fire first. Don't enable wdog1 unless ts7100-wdt
is disabled or is given a different timeout.